requirements.txt
# ─── ReelGenerator Dependencies ──────────────────────────────────────────────
# Core
requests>=2.31.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0

# Text-to-Speech (free, no API key)
edge-tts>=6.1.0

# Video assembly
moviepy>=2.0.0

# FFmpeg (required by moviepy — install separately)
# Windows: choco install ffmpeg  OR  download from https://ffmpeg.org/download.html
# Mac: brew install ffmpeg
# Linux: sudo apt install ffmpeg
